Separated patcher by its queue management and the active part.
